Refactor to support integration testing
In the production code, the handlers are very tied to the Axum
framework. As a result, it's more challenging to directly test
the handlers from integration tests. I started to refactor the
handlers so that an Axum handler delegates to a method that just
takes application types. I started with the
poker_session::create_session() method which now delegates to
poker_session::do_create_session().

Then, in the integration tests, I added a test to call
do_create_session().

One additional change that was required is to abstract the db connection
handling, since the prod code uses a connection pool but that's not
required in the test code. So added a new trait and changed the TestDb
type to implement that and just return an regular (not pooled)
connection.

Since the non-Axum handlers should be independent of Axum and things
like http status codes, these new delegates will use custom error types
and the Axum handlers will map those to the proper http status codes.
Added thiserror crate to help with these new app errors.